feat: tier-2 moat — framework /drop parity, dev inspector, @neodrag/test
- @neodrag/react|vue|solid/drop: sortable, drop plugins, sortableItem helpers
- @neodrag/test package (packages/testing) for Vitest mouse utilities
- @neodrag/core/dev: installInteractionInspector + debugSnapshot on engine
- Fix @neodrag/vanilla publishConfig for drop/plugins exports
- Expand Svelte drop exports; document drop/dev paths in framework docs
