+++ docs/src/content/framework/01-react/+page.mdx
@@ -8,7 +8,7 @@ tagline: 'A lightweight react hook to make your elements draggable.'
 npm i @neodrag/react@next
 ```
 
-Import built-in plugins from `@neodrag/core/plugins` or `@neodrag/react/plugins`. Use `useDroppable(ref, plugins)` for drop targets (same reactive `plugins` / `() => plugins` shape as `useDraggable`).
+Import built-in plugins from `@neodrag/core/plugins` or `@neodrag/react/plugins`. Drop targets and sortable lists use `@neodrag/react/drop` (`useDroppable`, `sortable`, `useSortableItem`). Same reactive `plugins` / `() => plugins` shape as `useDraggable`.
 
 ## Usage
