+++ docs/src/content/framework/02-vue/+page.mdx
@@ -8,7 +8,7 @@ tagline: 'A lightweight directive to make your elements draggable.'
 npm i @neodrag/vue@next
 ```
 
-Import built-in plugins from `@neodrag/core/plugins` or `@neodrag/vue/plugins`. Use `v-droppable` with the same reactive plugin input as `v-draggable`.
+Import built-in plugins from `@neodrag/core/plugins` or `@neodrag/vue/plugins`. Drop and sortable helpers are under `@neodrag/vue/drop` (`sortable`, `sortableItemPlugins`). Use `v-droppable` with the same reactive plugin input as `v-draggable`.
 
 ## Usage
