+++ docs/src/content/framework/03-solid/+page.mdx
@@ -8,6 +8,8 @@ tagline: 'A lightweight directive to make your elements draggable.'
 npm i @neodrag/solid@next
 ```
 
+Import built-in plugins from `@neodrag/core/plugins` or `@neodrag/solid/plugins`. Drop and sortable helpers live under `@neodrag/solid/drop` (`useDroppable`, `sortable`, `useSortableItem`).
+
 ## Usage
 
 Basic usage
