+++ docs/src/content/framework/05-core/+page.mdx
@@ -79,7 +79,7 @@ Destroys all bindings and listeners on this engine.
 
 ## Plugins
 
-Import built-in drag plugins from `@neodrag/core/plugins`. Drop helpers from `@neodrag/core/drop/plugins`; sortable from `@neodrag/core/drop`.
+Import built-in drag plugins from `@neodrag/core/plugins`. Drop helpers from `@neodrag/core/drop/plugins`; sortable from `@neodrag/core/drop`. In development, `installInteractionInspector(engine)` from `@neodrag/core/dev` overlays live drag/drop stats when `engine.dev` is true. Browser tests can use `@neodrag/test` (also re-exported from `@neodrag/core/testing`).
 
 The main `@neodrag/core` entry exports the engine, types, and default plugin presets only — optional plugins are not pulled in unless you import them.
