+++ packages/core/package.json
@@ -31,6 +31,10 @@
 			"types": "./src/testing/index.ts",
 			"import": "./src/testing/index.ts"
 		},
+		"./dev": {
+			"types": "./src/dev/index.ts",
+			"import": "./src/dev/index.ts"
+		},
 		"./package.json": "./package.json"
 	},
 	"publishConfig": {
@@ -60,6 +64,10 @@
 				"types": "./dist/testing/index.d.ts",
 				"import": "./dist/testing/index.js"
 			},
+			"./dev": {
+				"types": "./dist/dev/index.d.ts",
+				"import": "./dist/dev/index.js"
+			},
 			"./package.json": "./package.json"
 		}
 	},
@@ -72,7 +80,7 @@
 		"test:webkit": "vitest run --browser=webkit",
 		"test:firefox": "vitest run --browser=firefox",
 		"test:chromium": "vitest run --browser=chromium",
-		"test:unit": "vitest run tests/interactions/state-machine.test.ts tests/interactions/session.test.ts tests/interactions/session-lifecycle.test.ts tests/interactions/report-storage.test.ts tests/interactions/plugins-grid-controls.test.ts tests/interactions/apply-transform.test.ts tests/interactions/update.test.ts tests/interactions/update-pending.test.ts tests/interactions/drop-update.test.ts tests/interactions/plugin-slots.test.ts tests/interactions/plugins-phases.test.ts tests/interactions/coalescing.test.ts tests/interactions/drop-raf.test.ts tests/interactions/transform-update.test.ts tests/interactions/update-defaults.test.ts --browser.enabled=false",
+		"test:unit": "vitest run tests/interactions/state-machine.test.ts tests/interactions/session.test.ts tests/interactions/session-lifecycle.test.ts tests/interactions/report-storage.test.ts tests/interactions/plugins-grid-controls.test.ts tests/interactions/apply-transform.test.ts tests/interactions/update.test.ts tests/interactions/update-pending.test.ts tests/interactions/drop-update.test.ts tests/interactions/plugin-slots.test.ts tests/interactions/plugins-phases.test.ts tests/interactions/coalescing.test.ts tests/interactions/drop-raf.test.ts tests/interactions/transform-update.test.ts tests/interactions/update-defaults.test.ts tests/interactions/debug-snapshot.test.ts --browser.enabled=false",
 		"test:interactions": "vitest run --config vitest.interactions.config.ts",
 		"test:interactions:unit": "pnpm run test:unit",
 		"test:bench:guard": "vitest run --config vitest.bench.browser.config.ts benchmarks/browser/",
@@ -91,6 +99,7 @@
 	"author": "Puru Vijay",
 	"license": "MIT",
 	"devDependencies": {
+		"@neodrag/test": "workspace:*",
 		"@neodrag/svelte": "workspace:*",
 		"@vitest/coverage-v8": "3.2.4",
 		"jsdom": "^29.1.1",
