+++ packages/core/src/dev/index.ts
@@ -1,0 +1,7 @@
+export {
+	createInteractionInspector,
+	installInteractionInspector,
+	type InteractionInspector,
+	type InteractionInspectorOptions,
+} from './inspector.ts';
+export type { NeodragDebugSnapshot } from '../engine.ts';
