+++ packages/core/src/testing/index.ts
@@ -1,20 +1,1 @@
-export {
-	clickElement,
-	dragAndDrop,
-	dragUsingKeyboard,
-	getCursorPosition,
-	getCursorPositionRelativeToElement,
-	getElementCoords,
-	isCursorOverElement,
-	mouseClick,
-	mouseDoubleClick,
-	mouseDown,
-	mouseMove,
-	mouseUp,
-	mouseWheel,
-	startCursorTracking,
-	stopCursorTracking,
-	type MouseButton,
-	type MouseMoveOptions,
-	type MouseOptions,
-} from '../../tests/mouse.ts';
+export * from '@neodrag/test';
