+++ packages/react/package.json
@@ -18,6 +18,11 @@
 			"import": "./src/plugins.ts",
 			"default": "./src/plugins.ts"
 		},
+		"./drop": {
+			"types": "./src/drop/index.ts",
+			"import": "./src/drop/index.ts",
+			"default": "./src/drop/index.ts"
+		},
 		"./package.json": "./package.json"
 	},
 	"publishConfig": {
@@ -31,6 +36,11 @@
 				"types": "./dist/plugins.d.ts",
 				"import": "./dist/plugins.js",
 				"default": "./dist/plugins.js"
+			},
+			"./drop": {
+				"types": "./dist/drop/index.d.ts",
+				"import": "./dist/drop/index.js",
+				"default": "./dist/drop/index.js"
 			},
 			"./package.json": "./package.json"
 		}
