+++ packages/solid/src/drop/index.ts
@@ -1,0 +1,13 @@
+export {
+	accepts,
+	applySortableReorder,
+	highlight,
+	onDrop,
+	sortable,
+	type SortableMode,
+	type SortableOptions,
+	type SortablePreviewMeta,
+	type SortableReorderMeta,
+	type SortableStrategy,
+} from '@neodrag/core/drop';
+export { sortableItemPlugins, useSortableItem } from './sortable.ts';
