+++ packages/testing/package.json
@@ -1,0 +1,48 @@
+{
+	"name": "@neodrag/test",
+	"version": "3.0.0-next.9",
+	"type": "module",
+	"description": "Browser test helpers for Neodrag (Vitest + Playwright)",
+	"files": ["dist/*"],
+	"sideEffects": false,
+	"exports": {
+		".": {
+			"types": "./src/index.ts",
+			"import": "./src/index.ts"
+		},
+		"./package.json": "./package.json"
+	},
+	"publishConfig": {
+		"access": "public",
+		"exports": {
+			".": {
+				"types": "./dist/index.d.ts",
+				"import": "./dist/index.js"
+			},
+			"./package.json": "./package.json"
+		}
+	},
+	"scripts": {
+		"compile": "tsdown",
+		"prepublishOnly": "pnpm run compile"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/PuruVJ/neodrag.git"
+	},
+	"author": "Puru Vijay",
+	"license": "MIT",
+	"peerDependencies": {
+		"@vitest/browser": "^3.2.0",
+		"vitest": "^3.2.0"
+	},
+	"peerDependenciesMeta": {
+		"@vitest/browser": { "optional": true },
+		"vitest": { "optional": true }
+	},
+	"devDependencies": {
+		"@vitest/browser": "3.2.4",
+		"tsdown": "catalog:",
+		"vitest": "3.2.4"
+	}
+}
