+++ packages/testing/tsconfig.json
@@ -1,0 +1,13 @@
+{
+	"compilerOptions": {
+		"module": "ES2022",
+		"target": "ESNext",
+		"noEmit": true,
+		"strict": true,
+		"lib": ["ES2023.Array", "DOM"],
+		"moduleResolution": "bundler",
+		"allowImportingTsExtensions": true,
+		"types": ["@vitest/browser/providers/playwright"]
+	},
+	"include": ["./src/**/*.ts"]
+}
