+++ packages/testing/tsdown.config.ts
@@ -1,15 +1,10 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-	entry: {
-		index: './src/index.ts',
-		plugins: './src/plugins.ts',
-		'drop/index': './src/drop/index.ts',
-	},
+	entry: { index: './src/index.ts' },
 	format: 'esm',
 	dts: { resolve: true },
 	clean: true,
 	platform: 'browser',
 	treeshake: { moduleSideEffects: false },
-	deps: { neverBundle: ['react', 'react-dom', '@neodrag/core'] },
 });
