+++ packages/vanilla/src/drop.ts
@@ -1,3 +1,11 @@
 export { Neodrag, Droppable } from './index';
 export * from '@neodrag/core/drop/plugins';
-export { sortable, type SortableOptions, type SortableStrategy } from '@neodrag/core/drop';
+export {
+	sortable,
+	applySortableReorder,
+	type SortableOptions,
+	type SortableStrategy,
+	type SortableMode,
+	type SortablePreviewMeta,
+	type SortableReorderMeta,
+} from '@neodrag/core/drop';
