+++ packages/vue/tsdown.config.ts
@@ -1,7 +1,11 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-	entry: ['./src/index.ts'],
+	entry: {
+		index: './src/index.ts',
+		plugins: './src/plugins.ts',
+		'drop/index': './src/drop/index.ts',
+	},
 	format: 'esm',
 	dts: true,
 	clean: true,
